# Optimize FeatureDirectionalStatistics via LUT