Hey Cris,

I believe this patch is sane, but I don't have the full domain expertise. Do you agree with my assessment that there will be usually only a low number of different grey values encountered? This patch only makes sense in such a case, otherwise we would end up just doing more work than before.

At least in my adhoc tests, the assumption was valid. Before this patch I saw a big hotspot attributed to the repeated `sincos` calls. Afterwards only a bit of `std::unordered_map` lookups was left, but overall at a much reduced cost.

That said - are there some benchmarks for this code in the diplib repository itself that I could use to give you some hard numbers?