@crisluengo 2021-04-15
This is probably a good idea. I’d like to benchmark this with a floating-point input image too. Not sure it necessarily improves things there. Maybe we need to make it something for integer inputs only?