@crisluengo 2021-04-16
I wrote a simple comparison, timing the updated vs the old function. Please double-check my work, I might be missing something:
```cpp
#include "diplib.h"
#include "diplib/generation.h"
#include "diplib/segmentation.h"
#include "diplib/measurement.h"
#include "diplib/testing.h"

int main() {
   dip::Random random;
   // Create a test image
   dip::Image input({2048,2048}, 1, dip::DT_UINT8);
   dip::UniformNoise(input, input, random, 0, 256);
   dip::Image label = dip::Superpixels(input);
   // Measure
   dip::MeasurementTool measurementTool;
   dip::testing::Timer timer;
   auto msr = measurementTool.Measure( label, input, { "DirectionalStatistics" } );
   timer.Stop();
   std::cout << "DirectionalStatistics: " << timer << '\n';
   timer.Reset();
   msr = measurementTool.Measure( label, input, { "DirectionalStatisticsOriginal" } );
   timer.Stop();
   std::cout << "DirectionalStatisticsOriginal: " << timer << '\n';
   // Create float test image
   input.Convert(dip::DT_SFLOAT);
   dip::UniformNoise(input, input, random, 0, 2 * dip::pi);
   // Measure
   timer.Reset();
   msr = measurementTool.Measure( label, input, { "DirectionalStatistics" } );
   timer.Stop();
   std::cout << "DirectionalStatistics: " << timer << '\n';
   timer.Reset();
   msr = measurementTool.Measure( label, input, { "DirectionalStatisticsOriginal" } );
   timer.Stop();
   std::cout << "DirectionalStatisticsOriginal: " << timer << '\n';
}
```
So this creates a 2048x2048 image with 256 different values, and a labeled image with whatever random regions. It times the computation of the feature using the old and new versions of the code. It then fills the image with random floats in the same range (so that there are many more different values), and times the same computations.

In the first case I see 256 elements added to the `lut_`, and in the second case 3372595.

The output of the code is:
```none
DirectionalStatistics: elapsed time = 0.183064 s (wall), 0.180403 s (CPU)
DirectionalStatisticsOriginal: elapsed time = 0.140341 s (wall), 0.139095 s (CPU)
DirectionalStatistics: elapsed time = 2.41194 s (wall), 2.39287 s (CPU)
DirectionalStatisticsOriginal: elapsed time = 0.146867 s (wall), 0.14264 s (CPU)
```
So even for the integer image this looks like it is not really worth the effort, maybe with a much larger image it is? But for a floating-point image it certainly is bad.

The problem is that we don't have any functionality in the measurement infrastructure to use different code for an integer vs a floating-point input image.