@milianw 2021-04-19
Thanks for that benchmark. On my machine:

Without this patch:
```
DirectionalStatistics: elapsed time = 0.130948 s (wall), 0.130778 s (CPU)
DirectionalStatistics: elapsed time = 0.131066 s (wall), 0.346593 s (CPU)
```

With this patch:
```
DirectionalStatistics: elapsed time = 0.118109 s (wall), 0.117997 s (CPU)
DirectionalStatistics: elapsed time = 1.45075 s (wall), 1.63491 s (CPU)
```

The gain is miniscule and the potential for breakage large. I'll abandon this patch - it's not worth it.

On a somewhat related note though: Considering how often associative containers (i.e. now hash maps) are leveraged in diplib, you should really think about integrating a faster variant of that. I can highly recommend https://github.com/Tessil/robin-map e.g. which is liberally licensed and trivial to incorporate into a code base. The performance improvement over `std::unordered_map` is often immense.

Cheers