@crisluengo 2021-04-19
@milianw Thanks for that idea! I’m going to look into how to incorporate that into DIPlib, I’m sure it’ll be worth it.