@crisluengo 2021-04-21
@milianw See here: https://github.com/DIPlib/diplib/commit/8b1b3da0b34b8f6430e7c5f94d66d1a2b6451e93